requirements.txt
# FastAPI and ASGI server
fastapi>=0.100.0
uvicorn[standard]>=0.20.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.10.0

# Authentication and Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
python-multipart>=0.0.6

# Email
python-dotenv>=1.0.0

# HTTP requests for social auth
requests>=2.28.0

# Pydantic for data validation
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0

# Testing
pytest>=7.0.0
httpx>=0.24.0 